update, add full features

+++ src/controllers/Necessary/necessary-package.js
@@ -1,14 +1,20 @@
 const NecessaryPackage = require('../../models/Necessary/necessary-package');
 const {StatusCodes} = require('http-status-codes');
+const {BadRequestError, UnauthorizedError, NotFoundError} = require('../../errors');
+const Transaction = require('../../models/Activity/transaction');
 
 //[POST] /necessaryPackages/create
 const createNecessaryPackage = async (req, res) => {
-    const necessaryPackage = new NecessaryPackage(req.body);
-    await necessaryPackage.save();
-    res.status(StatusCodes.CREATED).json(necessaryPackage);    
+    if (req.priority === 1) {
+        const necessaryPackage = new NecessaryPackage(req.body);
+        await necessaryPackage.save();
+        res.status(StatusCodes.CREATED).json(necessaryPackage);    
+    } else {
+        throw new UnauthorizedError('You are not manager');
+    }
 };
 
-//[GET] /necessaryPackages/get
+//[GET] /necessaryPackages/getAll
 const getNecessaryPackageList =  async (req, res) => {         
     const necessaryPackages = await NecessaryPackage.find({}).populate('necessaries.necessary');
     if (!necessaryPackages) {
@@ -17,7 +23,7 @@ const getNecessaryPackageList =  async (req, res) => {
     res.status(StatusCodes.OK).json(necessaryPackages);        
 }
 
-//[GET] /necessaryPackages/get/:name
+//[GET] /necessaryPackages/:name/get
 const getNecessaryPackageByName =  async (req, res) => {
     const name = req.params.name;       
     const necessaryPackage = await NecessaryPackage.findOne({name}).populate('necessaries.necessary');
@@ -27,42 +33,68 @@ const getNecessaryPackageByName =  async (req, res) => {
     res.status(StatusCodes.OK).json(necessaryPackage);        
 }
 
-//[PATCH] /necessaryPackages/update/:name
+//[PATCH] /necessaryPackages/:name/update
 const updateNecessaryPackage = async (req, res) => {
-    const updates = Object.keys(req.body);
-    const allowedUpdates = ['name', 'necessaries', 'limitQuantityOfNecessary', 'limitQuantityOfPackageOverTime', 'limitTime'];
-    const isValidOperation = updates.every((update) => allowedUpdates.includes(update));
+    if (req.priority === 1) {
+        const updates = Object.keys(req.body);
+        const allowedUpdates = ['name', 'necessaries', 'limitQuantityOfNecessary', 'limitQuantityOfPackageOverTime', 'limitTime'];
+        const isValidOperation = updates.every((update) => allowedUpdates.includes(update));
 
-    if (!isValidOperation) {
-        throw new BadRequestError("Invalid update operation")
+        if (!isValidOperation) {
+            throw new BadRequestError("Invalid update operation")
+        }
+
+        const name = req.params.name; 
+        const necessaryPackage = await NecessaryPackage.findOne({name})                
+        if (!necessaryPackage) {
+            throw new NotFoundError("Necessary package not found");
+        }
+
+        updates.forEach((update) => necessaryPackage[update] = req.body[update]);
+        await necessaryPackage.save();
+
+        res.status(StatusCodes.OK).json(necessaryPackage);   
+    } else {
+        throw new UnauthorizedError('You are not manager');
     }
-
-    const name = req.params.name; 
-    const necessaryPackage = await NecessaryPackage.findOne({name})                
-    if (!necessaryPackage) {
-        throw new NotFoundError("Necessary package not found");
-    }
-
-    updates.forEach((update) => necessaryPackage[update] = req.body[update]);
-    await necessaryPackage.save();
-
-    res.status(StatusCodes.OK).json(necessaryPackage);    
 };
 
-//[DELETE] /necessaryPackage/delete/:name
+//[DELETE] /necessaryPackage/:name/delete
 const deleteNecessaryPackageByName =  async (req, res) => {  
-    const name = req.params.name;  
-    const necessaryPackage = await NecessaryPackage.findOneAndDelete({name});
-    if (!necessaryPackage) {
-        throw new NotFoundError("Necessary package not found");
+    if (req.priority === 1) {
+        const name = req.params.name;  
+        const necessaryPackage = await NecessaryPackage.findOneAndDelete({name});
+        if (!necessaryPackage) {
+            throw new NotFoundError("Necessary package not found");
+        }
+        res.status(StatusCodes.OK).json(necessaryPackage);
+    } else {
+        throw new UnauthorizedError('You are not manager');
     }
-    res.status(StatusCodes.OK).json(necessaryPackage);      
 }
+
+//[POST] /necessaryPackages/buy
+const buyNecessaryPackages = async (req, res) => {
+    //update transaction history
+    const necessaryPackages = req.body.necessaryPackages;
+    const userId = req.user._id;
+    const transaction = new Transaction({
+        user:userId,
+        necessaryPackages
+    })
+    //update debt of user
+    const savedTransaction = await transaction.save();        
+    req.user.debt += savedTransaction.totalPrice;
+    await req.user.save();
+
+    res.status(StatusCodes.OK).json(transaction);    
+};
 
 module.exports = {
     createNecessaryPackage,
     getNecessaryPackageList,
     getNecessaryPackageByName,
     updateNecessaryPackage,
-    deleteNecessaryPackageByName
+    deleteNecessaryPackageByName,
+    buyNecessaryPackages
 };
